# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Test Academy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  admin: 'admin',
  student: 'student',
} as const;

export interface AuthUser {
  /** @nullable */
  email?: string | null;
  role: AuthUserRole;
  /** @nullable */
  name?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface StudentLoginInput {
  /** @minLength 1 */
  name: string;
}

export interface EligibilityItem {
  id: number;
  category: string;
  title: string;
  content: string;
  isPublished: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface EligibilityInput {
  category: string;
  title: string;
  content: string;
  isPublished?: boolean;
}

export interface EligibilityUpdate {
  category?: string;
  title?: string;
  content?: string;
}

export interface McqLevel {
  id: number;
  levelNumber: number;
  title: string;
  questionCount: number;
}

export interface McqQuestion {
  id: number;
  levelId: number;
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: string;
  orderIndex: number;
}

export type McqQuestionInputCorrectAnswer = typeof McqQuestionInputCorrectAnswer[keyof typeof McqQuestionInputCorrectAnswer];


export const McqQuestionInputCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface McqQuestionInput {
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: McqQuestionInputCorrectAnswer;
  orderIndex?: number;
}

export type McqQuestionUpdateCorrectAnswer = typeof McqQuestionUpdateCorrectAnswer[keyof typeof McqQuestionUpdateCorrectAnswer];


export const McqQuestionUpdateCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface McqQuestionUpdate {
  question?: string;
  optionA?: string;
  optionB?: string;
  optionC?: string;
  optionD?: string;
  correctAnswer?: McqQuestionUpdateCorrectAnswer;
  orderIndex?: number;
}

export type McqQuestionCorrectAnswer = typeof McqQuestionCorrectAnswer[keyof typeof McqQuestionCorrectAnswer];


export const McqQuestionCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export type McqAnswerSelectedAnswer = typeof McqAnswerSelectedAnswer[keyof typeof McqAnswerSelectedAnswer];


export const McqAnswerSelectedAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface McqAnswer {
  questionId: number;
  selectedAnswer: McqAnswerSelectedAnswer;
}

export interface McqSubmission {
  levelId: number;
  answers: McqAnswer[];
}

export interface McqResult {
  score: number;
  total: number;
  percentage: number;
}

export interface EssayItem {
  id: number;
  title: string;
  content: string;
  isPublished: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface EssayInput {
  title: string;
  content: string;
  isPublished?: boolean;
}

export interface EssayUpdate {
  title?: string;
  content?: string;
}

export interface PhysicalItem {
  id: number;
  category: string;
  title: string;
  content: string;
  isPublished: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface PhysicalInput {
  category: string;
  title: string;
  content: string;
  isPublished?: boolean;
}

export interface PhysicalUpdate {
  category?: string;
  title?: string;
  content?: string;
}

export type FinalTestQuestionCorrectAnswer = typeof FinalTestQuestionCorrectAnswer[keyof typeof FinalTestQuestionCorrectAnswer];


export const FinalTestQuestionCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface FinalTestQuestion {
  id: number;
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: FinalTestQuestionCorrectAnswer;
  orderIndex: number;
}

export type FinalTestQuestionInputCorrectAnswer = typeof FinalTestQuestionInputCorrectAnswer[keyof typeof FinalTestQuestionInputCorrectAnswer];


export const FinalTestQuestionInputCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface FinalTestQuestionInput {
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: FinalTestQuestionInputCorrectAnswer;
  orderIndex?: number;
}

export type FinalTestQuestionUpdateCorrectAnswer = typeof FinalTestQuestionUpdateCorrectAnswer[keyof typeof FinalTestQuestionUpdateCorrectAnswer];


export const FinalTestQuestionUpdateCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface FinalTestQuestionUpdate {
  question?: string;
  optionA?: string;
  optionB?: string;
  optionC?: string;
  optionD?: string;
  correctAnswer?: FinalTestQuestionUpdateCorrectAnswer;
  orderIndex?: number;
}

export type FinalTestAnswerSelectedAnswer = typeof FinalTestAnswerSelectedAnswer[keyof typeof FinalTestAnswerSelectedAnswer];


export const FinalTestAnswerSelectedAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface FinalTestAnswer {
  questionId: number;
  selectedAnswer: FinalTestAnswerSelectedAnswer;
}

export interface FinalTestSubmission {
  studentName: string;
  answers: FinalTestAnswer[];
}

export interface FinalTestResult {
  id: number;
  studentName: string;
  score: number;
  total: number;
  percentage: number;
  submittedAt: string;
}

export interface PublishInput {
  isPublished: boolean;
}

export interface DashboardStats {
  totalStudents: number;
  activeStudents: number;
  publishedContent: number;
  totalLevels: number;
  eligibilityCount: number;
  mcqLevelsWithQuestions: number;
  essayCount: number;
  physicalCount: number;
  finalTestQuestionCount: number;
  finalTestResultCount: number;
}

export type ListEligibilityParams = {
/**
 * Set to "true" for admin to see all items including unpublished
 */
all?: string;
};

export type ListEssaysParams = {
all?: string;
};

export type ListPhysicalParams = {
all?: string;
category?: string;
};

export type ListFinalTestQuestionsParams = {
all?: string;
};