# lib/api-zod/src/generated/types/eligibilityUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Test Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface EligibilityUpdate {
  category?: string;
  title?: string;
  content?: string;
}