# lib/api-zod/src/generated/types/essayInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Test Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface EssayInput {
  title: string;
  content: string;
  isPublished?: boolean;
}