# lib/api-zod/src/generated/types/finalTestAnswerSelectedAnswer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Test Academy API
 * OpenAPI spec version: 0.1.0
 */

export type FinalTestAnswerSelectedAnswer = typeof FinalTestAnswerSelectedAnswer[keyof typeof FinalTestAnswerSelectedAnswer];


export const FinalTestAnswerSelectedAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;