# lib/api-zod/src/generated/types/finalTestQuestionCorrectAnswer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Test Academy API
 * OpenAPI spec version: 0.1.0
 */

export type FinalTestQuestionCorrectAnswer = typeof FinalTestQuestionCorrectAnswer[keyof typeof FinalTestQuestionCorrectAnswer];


export const FinalTestQuestionCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;